requirements.txt
antiorm
asgiref
attrs
bcrypt
blinker
cffi
click
colorama
contourpy
contrib
crispy-bootstrap5
cryptography
cycler
DateTime
db
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
environ
Flask
Flask-Bcrypt
Flask-MySQL
Flask-MySQLdb
fonttools
gunicorn
h11
itsdangerous
Jinja2
jmespath
jsonschema
jsonschema-specifications
kiwisolver
Markdown
MarkupSafe
matplotlib
MouseInfo
mysql-connector-python
mysqlclient
numpy
opencv-contrib-python
packaging
pillow
psycopg2
PyAutoGUI
pycparser
PyGetWindow
PyMsgBox
PyMySQL
pyparsing
pyperclip
PyRect
PyScreeze
python-dateutil
pytweening
pytz
PyYAML
referencing
rpds-py
shortcuts
six
sqlparse
template
toml
typing_extensions
tzdata
utils
uvicorn
views
Werkzeug
zope.interface